requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
